chore: init screen
