--- app/src/main/java/com/example/mvvm/MainApplication.kt
@@ -1,8 +1,0 @@
-package com.example.mvvm
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class MainApplication : Application() {
-}
