--- app/src/main/java/com/example/mvvm/repositories/MainLog.kt
@@ -1,7 +1,0 @@
-package com.example.mvvm.repositories
-
-interface MainLog {
-   fun d(tag: String, msg: String)
-   fun i(tag: String, msg: String)
-   fun e(tag: String, msg: String)
-}
