--- app/src/main/java/com/example/mvvm/repositories/Store.kt
@@ -1,6 +1,0 @@
-package com.example.mvvm.repositories
-
-interface Store {
-    fun getValue(key: String): String
-    fun setValue(key: String, value: String)
-}
