--- app/src/main/java/com/example/mvvm/ui/screen/detail/DetailScreen.kt
@@ -1,13 +1,0 @@
-package com.example.mvvm.ui.screen.detail
-
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
-
-@Composable
-fun DetailScreen(navController: NavHostController, hiltViewModel: DetailViewModel) {
-    TextButton(onClick = { navController.popBackStack() }) {
-        Text(text = "Back")
-    }
-}
