+++ app/src/main/java/com/example/nocket/MainApplication.kt
@@ -1,0 +1,8 @@
+package com.example.nocket
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MainApplication : Application() {
+}
