+++ app/src/main/java/com/example/nocket/common/enum/LoadStatus.kt
@@ -1,0 +1,8 @@
+package org.com.hcmurs.common.enum
+
+sealed class LoadStatus(val description: String = "") {
+    class Init(): LoadStatus()
+    class Loading() : LoadStatus()
+    class Success(): LoadStatus()
+    class Error(val error: String): LoadStatus(error)
+}
