+++ app/src/main/java/com/example/nocket/components/common/CommonTopBar.kt
@@ -1,0 +1,62 @@
+package com.example.nocket.components.common
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.navigation.NavController
+import com.example.nocket.Screen
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CommonTopBar(
+    navController: NavController,
+    title: String,
+    actions: @Composable RowScope.() -> Unit = {},
+    bottomContent: @Composable (() -> Unit)? = null
+) {
+    Column {
+        CenterAlignedTopAppBar(
+            title = {
+                Box(
+                    modifier = Modifier.fillMaxWidth(),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Text(
+                        text = title,
+                        color = Color.White,
+                        fontWeight = FontWeight.Normal,
+                    )
+                }
+            },
+            navigationIcon = {
+                IconButton(onClick = { navController.navigate(Screen.Home.route) }) {
+                    Icon(
+                        imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                        contentDescription = "Quay lại",
+                        tint = Color.White
+                    )
+                }
+            },
+            actions = actions,
+            colors = TopAppBarDefaults.topAppBarColors(
+                containerColor = Color.Blue
+            )
+        )
+
+        bottomContent?.invoke()
+    }
+}
