+++ app/src/main/java/com/example/nocket/constants/AuthConstants.kt
@@ -1,0 +1,6 @@
+package org.com.hcmurs.constant
+
+object AuthConstants {
+    const val SHARED_PREFS_NAME = "auth_prefs"
+    const val TOKEN_KEY = "access_token"
+}
