+++ app/src/main/java/com/example/nocket/constants/ScreenTitle.kt
@@ -1,0 +1,13 @@
+package com.example.nocket.constants
+
+import androidx.annotation.StringRes
+import com.example.nocket.R
+
+enum class ScreenTitle(@StringRes val titleRes: Int) {
+    HOME(R.string.home),
+    MESSAGE(R.string.message),
+    POST(R.string.post),
+    PROFILE(R.string.profile),
+    RELATIONSHIP(R.string.relationship),
+    SETTING(R.string.setting),
+}
