+++ app/src/main/java/com/example/nocket/preview/HomeScreenPreview.kt
@@ -1,0 +1,12 @@
+package com.example.nocket.preview
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.compose.rememberNavController
+import com.example.nocket.ui.screen.home.HomeScreen
+
+@Preview(showBackground = true)
+@Composable
+fun HomeScreenPreview() {
+    HomeScreen(rememberNavController())
+}
