+++ app/src/main/java/com/example/nocket/repositories/MainLog.kt
@@ -1,0 +1,7 @@
+package com.example.nocket.repositories
+
+interface MainLog {
+   fun d(tag: String, msg: String)
+   fun i(tag: String, msg: String)
+   fun e(tag: String, msg: String)
+}
