+++ app/src/main/java/com/example/nocket/repositories/MainLogImpl.kt
@@ -1,0 +1,18 @@
+package com.example.nocket.repositories
+
+import android.util.Log
+import javax.inject.Inject
+
+class MainLogImpl @Inject constructor() : MainLog {
+    override fun d(tag: String, msg: String) {
+        Log.d(tag, msg)
+    }
+
+    override fun i(tag: String, msg: String) {
+        Log.i(tag, msg)
+    }
+
+    override fun e(tag: String, msg: String) {
+        Log.e(tag, msg)
+    }
+}
