+++ app/src/main/java/com/example/nocket/repositories/Store.kt
@@ -1,0 +1,6 @@
+package com.example.nocket.repositories
+
+interface Store {
+    fun getValue(key: String): String
+    fun setValue(key: String, value: String)
+}
