+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,10 @@
 <resources>
-    <string name="app_name">MVVM</string>
+    <string name="app_name">nocket</string>
+
+    <string name="home">Home</string>
+    <string name="message">Message</string>
+    <string name="profile">Profile</string>
+    <string name="post">Post</string>
+    <string name="relationship">Relationship</string>
+    <string name="setting">Settings</string>
 </resources>
